Rename types to types_check
